+++ CHANGELOG.md
@@ -15,9 +15,11 @@ Polyverse Boost ReST API (Backend)
 ### Enhancements
 - Add delays (2-3 seconds) to OpenAI file uploads to avoid API throttling (60 calls/minute)
 - Delete OpenAI files when new resource data is uploaded - from /api/user_project/{org}/{project}/data_references (POST)
+- Ensure Project Status is refreshed after a File Upload has occured (e.g. /api/user_project/{org}/{project}/status (GET))
 
 ### Bug Fixes
 - Handle rename of last_updated -> lastUpdated in /user_project/{org}/{project}/data/{resource}/status (GET)
+- Fix Project Status issue incorrectly reporting resources not Uploaded to OpenAI
 
 ## Version 0.10.0: February 12th, 2024